Add JET static analysis tests and type annotations
- Add JET.jl as a test dependency
- Add JET test in test/jet.jl to verify package has no static analysis issues
- Add `::Integer` type annotation to `generate_priors` and `generate_theta` functions
- Add "JET" test group to runtests.jl

JET analysis confirms DiffEqBayes has 0 issues when filtering to target module only.
(All 388 issues in initial report were from dependencies like StanSample, SymbolicUtils, etc.)

+++ src/stan_inference.jl
@@ -10,7 +10,7 @@ end
 
 struct StanODEData end
 
-function generate_priors(n, priors)
+function generate_priors(n::Integer, priors)
     priors_string = ""
     if priors === nothing
         for i in 1:n
@@ -27,7 +27,7 @@ function generate_priors(n, priors)
     return priors_string
 end
 
-function generate_theta(n, priors)
+function generate_theta(n::Integer, priors)
     theta = ""
     for i in 1:n
         upper_bound = ""
